discord-kickbot/requirements.txt
discord.py
PyNaCl==1.3.0
dnspython==1.16.0
boto3